// src/domain/types.ts
export type EffectDto =
| { kind: "HCDistortion"; data: HcDistortionDto }
| { kind: "SCDistortion"; data: ScDistortionDto }
| { kind: "Delay"; data: DelayDto }
| { kind: "Cabinet"; data: CabinetDto }
| { kind: "Wah"; data: WahDto }